attempt to correct compiler error

+++ src/core/serializers/GmtlJsonSerializer.h
@@ -1,17 +1,16 @@
 #pragma once
 #include <gmtl/gmtl.h>
-
 #include <nlohmann/json.hpp>
 
 // serializers
 namespace gmtl 
 {
-template<class DATA_TYPE, size_t S>
+template<class DATA_TYPE, unsigned S>
 inline void to_json(nlohmann::json& j, const VecBase<DATA_TYPE, S> & p)
 {
     j = nlohmann::json::array({p.mData});
 }
-template <class DATA_TYPE, size_t S>
+template <class DATA_TYPE, unsigned S>
 inline void from_json(const nlohmann::json& j, VecBase<DATA_TYPE, S>& p)
 {
     j.get_to(p.mData);
